Run user command as a service
Instead of running the user-given command as a backround process of the init process, provide a `cmdline` service which executes the user-given command instead. This allows us to wait for all services to be "up" before we begin executing the user command.

Running the user command as a supervised service also gives more control to the init process.